requirements.txt
streamlit>=1.32.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
scikit-image>=0.21.0
